refactor: rename FunctionPerformance directory and page to reflect new naming conventions
