+++ client/src/pages/PerformanceOverview/AverageBilledDuration/AverageBilledDuration.tsx
